requirements.txt
Pillow
types-Pillow
webcolors
webcolors-stubs